Work on water flow and player model

+++ include/Events/Event.hpp
@@ -77,7 +77,7 @@ struct EnterChunkEvent {
 
 struct BlockChangedEvent {
     Application* app;
-    glm::vec3 position;
+    glm::ivec3 position;
     BlockId prevBlock;
     BlockId newBlock;
